Fix issue where parser tests failed because Mocha couldn't load a module
The parser tests still fail, but now in a good way, yay!

Mocha couldn't load the mi_output_parser module because I used
require('mi_output_parser') in the tests which meant that NodeJS went
looking for the module in node_modules instead of the src dir. The fix
consists of using a file path in the require() instead of just the
module name so that NodeJS looks for the module in the right place,
and also to make the module "external" as far as TypeScript is concerned
(or else the TypeScript compiler refuses to let you require the module
via a file path).