Emit events for target running and stopped notifications
There is only one type of running notification, but the stopped
notification has a bunch of different reasons for the stop, some of
which are emitted as additional events. Currently if a target stops due
to a breakpoint, signal, or exception EVENT_TARGET_STOPPED is emitted,
followed by a more specialized event that provides additional
information.