Rename DebugSession.getSelectedStackFrame() to getStackFrame()
This method can now be used to retrieve information about any frame on
the stack, not just the selected one... at least when using GDB, LLDB
unfortunately will just provide information about the selected frame,
but hopefully that will be fixed in the near future.

Also the --thread option for MI commands is not specific to LLDB as I
previously assumed, it's also supported by GDB, so doc comments have
been updated to reflect this.