Trap exceptions thrown during transformation of the MI command response
In most cases the response to an MI command needs to be transformed into
a more useful form, any exceptions thrown during the transformation
should be trapped and handed over to the promise within which the
command was created. This ensures that exceptions travel down the
promise chain as expected instead of immediately terminating the
program.