Prevent NPM from attempting to build test targets on install
The default NPM preinstall script will attempt to compile a native
module if it sees a binding.gyp in the module root directory. However,
currently I'm only using binding.gyp to build the test targets, which
are not needed unless you're going to run the tests. The tests should
only be run when developing the module itself, not when installing this
module as a dependency, so to prevent NPM from building the test targets
I've set the preinstall script to do nothing.