Add a reference to bunyan.d.ts in package.json
Some bunyan types are visible in the generated typings for this package,
so to make it easier to consume this package in other TypeScript
projects `bunyan.d.ts` is referenced in `package.json`. This makes it
possible to use the `tsd link` command to automatically reference the
bunyan.d.ts in this package within a TypeScript project that consumes
this package.

The reference to node.d.ts has been removed from bunyan.d.ts to avoid
causing conflicts in any project that consumes this package and also
references its own copy of node.d.ts.