Generate the .js files for tests into a separate test-js directory
This is to support the oddball development environment I setup for this
package, where the directories containing .ts files are shared between
Windows and an Ubuntu VM, but the generated .js files are not. Any code
that actually gets executed needs to reside on the machine that runs it,
otherwise Node ends up looking for modules on the wrong machine and all
hell breaks loose.